Update: changed comments and restructured
